Update implementation to use g:vars and partials.
This change attempts to add some structure to the code. Since markdown
has several syntax options for the same semantics, we leave the
specification of which syntax to use to user set variables. If these are
not set, defaults are used. This reduces the number of command mappings
required.

Additionally, we now use partials for function invocation. This
simplifies the API for those who may want to rebind commands or
shortcuts while maintaining the abstraction/code reuse in the
implementation.